back/requirements.txt
anthropic==0.72.0
apscheduler==3.11.1
fastapi==0.133.1
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
httpx==0.28.1
langchain==0.1.10
langchain-anthropic==0.1.13
langchain-core==0.1.53
langchain-openai==0.1.6
pydantic==2.12.5
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
uvicorn==0.41.0
dateparser==1.2.2
